feat(web): extract repo worktree and ticket read model services
- Add repo_worktree_read_models.py and ticket_read_models.py service modules
- Harden hub_repo_routes read models with typed contracts
- Update frontend readModelContracts, data loaders, and view models
- Fix page/ticket/worktree Svelte views for new read model shapes
- Add test coverage for read model routes and frontend loaders
