+++ src/codex_autorunner/surfaces/web/routes/hub_repo_routes/__init__.py
@@ -15,12 +15,13 @@ __all__ = [
     "build_hub_repo_read_model_router",
 ]
 
+from ...services.repo_worktree_read_models import RepoWorktreeReadModelService
 from .cache_coordinator import HubCacheCoordinator
 from .channels import build_hub_channel_router
 from .crud import build_hub_repo_crud_router
 from .destinations import HubDestinationService
 from .mount_manager import HubMountManager
-from .read_models import RepoWorktreeReadModelService, build_hub_repo_read_model_router
+from .read_models import build_hub_repo_read_model_router
 from .repo_listing import build_hub_repo_listing_router
 from .run_control import HubRunControlService
 from .services import HubRepoEnricher
