+++ src/codex_autorunner/web_frontend/src/routes/repos/[repoId]/+page.svelte
@@ -40,18 +40,18 @@
     const payload = snapshot as {
       identity: Record<string, unknown>;
       topology: Record<string, unknown>;
-      scopedRuns: Record<string, unknown>[];
-      scopedChats: Record<string, unknown>[];
-      scopedTickets: Record<string, unknown>[];
+      runQueue: Record<string, unknown>[];
+      chatQueue: Record<string, unknown>[];
+      ticketQueue: Record<string, unknown>[];
       contextspaceSummary: Record<string, unknown>[];
       currentArtifacts: Record<string, unknown>[];
     };
     const baseSource = {
       repos: [mapRepoSummary(payload.identity)],
       worktrees: childrenFromTopology(payload.topology).map(mapWorktreeSummary),
-      runs: payload.scopedRuns.map(mapPmaRunProgress),
-      chats: payload.scopedChats.map(mapPmaChatSummary),
-      tickets: payload.scopedTickets.map(mapTicketSummary),
+      runs: payload.runQueue.map(mapPmaRunProgress),
+      chats: payload.chatQueue.map(mapPmaChatSummary),
+      tickets: payload.ticketQueue.map(mapTicketSummary),
       contextspaceDocs: payload.contextspaceSummary.map(mapContextspaceDocument),
       artifacts: payload.currentArtifacts.map(mapSurfaceArtifact)
     };
