+++ src/codex_autorunner/web_frontend/src/routes/repos/[repoId]/load.test.ts
@@ -147,9 +147,9 @@ function repoDetailSnapshot(repoId = 'repo-1'): RepoWorktreeDetailSnapshot {
     parentLinks: {},
     topology: { children: [] },
     runtime: {},
-    scopedTickets: [],
-    scopedRuns: [],
-    scopedChats: [],
+    ticketQueue: [],
+    runQueue: [],
+    chatQueue: [],
     contextspaceSummary: [],
     currentArtifacts: [],
     ticketWindow: { limit: 20, totalEstimate: 0, totalIsExact: true },
@@ -171,9 +171,9 @@ function worktreeDetailSnapshot(worktreeId = 'wt-1'): RepoWorktreeDetailSnapshot
     parentLinks: {},
     topology: {},
     runtime: {},
-    scopedTickets: [],
-    scopedRuns: [],
-    scopedChats: [],
+    ticketQueue: [],
+    runQueue: [],
+    chatQueue: [],
     contextspaceSummary: [],
     currentArtifacts: [],
     ticketWindow: { limit: 20, totalEstimate: 0, totalIsExact: true },
