+++ src/codex_autorunner/web_frontend/src/routes/repos/[repoId]/tickets/[ticketId]/load.test.ts
@@ -170,6 +170,10 @@ function ticketDetailSnapshot(ticketId = 't-1', ownerId = 'repo-1'): TicketDetai
     artifacts: [],
     dispatchWindow: { limit: 20, totalEstimate: 0, totalIsExact: true },
     dispatches: [],
+    ticketDetail: {},
+    ticketQueue: [],
+    runQueue: [],
+    chatQueue: [],
     repair: repair('/hub/read-models/tickets/t-1')
   };
 }
