+++ src/codex_autorunner/web_frontend/src/routes/tickets/[ticketId]/load.test.ts
@@ -157,6 +157,10 @@ function ticketDetailSnapshot(ticketId = 't-1'): TicketDetailSnapshot {
     artifacts: [],
     dispatchWindow: { limit: 20, totalEstimate: 0, totalIsExact: true },
     dispatches: [],
+    ticketDetail: {},
+    ticketQueue: [],
+    runQueue: [],
+    chatQueue: [],
     repair: repair('/hub/read-models/tickets/t-1')
   };
 }
