+++ src/codex_autorunner/web_frontend/src/routes/worktrees/[worktreeId]/load.test.ts
@@ -145,9 +145,9 @@ function worktreeDetailSnapshot(worktreeId = 'wt-1'): RepoWorktreeDetailSnapshot
     parentLinks: {},
     topology: {},
     runtime: {},
-    scopedTickets: [],
-    scopedRuns: [],
-    scopedChats: [],
+    ticketQueue: [],
+    runQueue: [],
+    chatQueue: [],
     contextspaceSummary: [],
     currentArtifacts: [],
     ticketWindow: { limit: 20, totalEstimate: 0, totalIsExact: true },
